[Feat] amplitude event data 조회 로직 추가 (#586)
* feat: amplitude raw data 테이블에서 조회하도록 추가

- 최소한의 필요한 컬럼만 orm으로 매핑
- @Immutable 추가하여 read-only entity임을 명시

related issue #585

* chore: 기존 amplitude api 응답 처리 로직 삭제

related issue #585

* chore: 불필요한 print문 제거

related issue #585

* feat: crew event data 반영

related issue #585

* feat: 플그유형 coffee chat 계산 기준 변경

- 매직 넘버 상수화

related issue #585

* fix: long -> double

- 정수형 나눗셈을 수행하지 않도록 수정

related issue #585

* chore: 주석 추가

related issue #585

* chore: print문 제거

related issue #585

* chore: coffee chat 시작 기준일자 변경

related issue #585

+++ src/main/java/org/sopt/makers/internal/report/dto/PlayGroundTypeStatsDto.java
@@ -6,16 +6,16 @@ import java.util.List;
 import org.sopt.makers.internal.report.domain.PlaygroundType;
 
 public record PlayGroundTypeStatsDto(
-	Long community,
-	Long member,
-	Long project,
-	Long wordChainGame,
-	Long coffeeChat,
-	Long crew
+	double community,
+	double member,
+	double project,
+	double wordChainGame,
+	double coffeeChat,
+	double crew
 ) {
 	public PlaygroundType getTopStats() {
-		List<Long> values = Arrays.asList(community, member, project, wordChainGame, coffeeChat, crew);
-		long max = values.stream().max(Long::compareTo).orElse(0L);
+		List<Double> values = Arrays.asList(community, member, project, wordChainGame, coffeeChat, crew);
+		double max = values.stream().max(Double::compareTo).orElse(0.0);
 
 		if (max == community) {
 			return PlaygroundType.COMMUNITY;
